requirements.txt
accelerate>=0.33.0
transformers>=4.44.0
datasets>=2.21.0
evaluate>=0.4.0
torch>=2.2.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
pandas>=2.0.0
yt-dlp>=2024.3.10
audio-separator>=0.18.0
onnxruntime>=1.15.0
fastapi>=0.110.0
uvicorn>=0.29.0
pypdf>=4.0.0
openai-whisper
python-multipart>=0.0.9
